api_yamdb/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2.18
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
gunicorn==20.0.4
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.1.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-dotenv==0.21.1
pytz==2023.3
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15